% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{as.data.frame.crazyforest.prediction}
\alias{as.data.frame.crazyforest.prediction}
\title{as.data.frame method for CrazyForest prediction}
\usage{
\method{as.data.frame}{crazyforest.prediction}(x, ...)
}
\arguments{
\item{x}{CrazyForest prediction object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Predictions as data frame.
}
\description{
as.data.frame method for CrazyForest prediction
}
\author{
Javier Pérez-Rodríguez
}
