% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance.crazyforest}
\alias{importance.crazyforest}
\alias{importance}
\title{crazyforest variable importance}
\usage{
\method{importance}{crazyforest}(x, ...)
}
\arguments{
\item{x}{crazyforest object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Variable importance measures.
}
\description{
Extract variable importance of crazyforest object.
}
\seealso{
\code{\link{crazyforest}}
}
\author{
Javier Pérez-Rodríguez
}
