% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predictions.crazyforest.prediction}
\alias{predictions.crazyforest.prediction}
\alias{predictions}
\title{CrazyForest predictions}
\usage{
\method{predictions}{crazyforest.prediction}(x, ...)
}
\arguments{
\item{x}{CrazyForest prediction object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Predictions: Classes for Classification forests, Numerical values for Regressions forests and the estimated survival functions for all individuals for Survival forests.
}
\description{
Extract predictions of CrazyForest prediction object.
}
\seealso{
\code{\link{crazyforest}}
}
\author{
Javier Pérez-Rodríguez
}
