% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.crazyforest.forest}
\alias{print.crazyforest.forest}
\title{Print CrazyForest forest}
\usage{
\method{print}{crazyforest.forest}(x, ...)
}
\arguments{
\item{x}{Object of class 'crazyforest.forest'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print contents of CrazyForest forest object.
}
\author{
Javier Pérez-Rodríguez
}
