% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.crazyforest.prediction}
\alias{print.crazyforest.prediction}
\title{Print CrazyForest prediction}
\usage{
\method{print}{crazyforest.prediction}(x, ...)
}
\arguments{
\item{x}{Object of class 'crazyforest.prediction'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print contents of CrazyForest prediction object.
}
\author{
Javier Pérez-Rodríguez
}
