% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.crazyforest}
\alias{print.crazyforest}
\title{Print CrazyForest}
\usage{
\method{print}{crazyforest}(x, ...)
}
\arguments{
\item{x}{Object of class 'crazyforest'.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print contents of CrazyForest object.
}
\seealso{
\code{\link{crazyforest}}
}
\author{
Javier Pérez-Rodríguez
}
