% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{str.crazyforest.forest}
\alias{str.crazyforest.forest}
\title{str method for CrazyForest forest}
\usage{
\method{str}{crazyforest.forest}(object, max.level = 2, ...)
}
\arguments{
\item{object}{Object of class 'crazyforest.forest'.}

\item{max.level}{Maximum level of nesting to display.}

\item{...}{Further arguments passed to \code{str}.}
}
\description{
Print structure of a CrazyForest forest object.
}
\keyword{internal}
