% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{str.crazyforest}
\alias{str.crazyforest}
\title{str method for CrazyForest}
\usage{
\method{str}{crazyforest}(object, max.level = 2, ...)
}
\arguments{
\item{object}{Object of class 'crazyforest'.}

\item{max.level}{Maximum level of nesting to display.}

\item{...}{Further arguments passed to \code{str}.}
}
\description{
Print structure of a CrazyForest object.
}
\keyword{internal}
