% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timepoints.R
\name{timepoints.crazyforest.prediction}
\alias{timepoints.crazyforest.prediction}
\title{CrazyForest timepoints}
\usage{
\method{timepoints}{crazyforest.prediction}(x, ...)
}
\arguments{
\item{x}{CrazyForest Survival prediction object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Unique death times
}
\description{
Extract unique death times of CrazyForest Survival prediction object.
}
\seealso{
\code{\link{crazyforest}}
}
\author{
Javier Pérez-Rodríguez
}
